pyproject.toml
[tool.poetry]
name = "advanceprofit-ea"
version = "0.1.0"
description = "ML-powered trading bot with XGBoost predictions for forex/crypto/stocks"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
# Core ML and Data Processing
xgboost = "^2.1.3"
pandas = "^2.3.3"
numpy = "^2.3.4"
scikit-learn = "^1.7.2"
joblib = "^1.5.2"

# Hyperparameter Optimization
optuna = "^3.6.2"
optuna-dashboard = "^0.16.1"

# Experiment Tracking
mlflow = "^2.19.0"

# Data Collection
yfinance = "^0.2.50"
ccxt = "^4.4.40"
MetaTrader5 = "^5.0.4563"

# Backtesting
backtrader = "^1.9.78"

# API Development
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.5"

# Visualization
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
plotly = "^5.24.1"

# Jupyter
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"

# Utilities
PyYAML = "^6.0.1"
requests = "^2.32.3"
tqdm = "^4.67.1"
ta = "^0.11.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"

# Code Quality
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.14.1"
isort = "^5.13.2"

# Development Tools
ipdb = "^0.13.13"
pre-commit = "^4.0.1"

[tool.poetry.scripts]
# Useful scripts you can run with `poetry run <command>`
train = "src.train:main"
serve = "src.api:start_server"
backtest = "src.backtest:main"
fetch-data = "data_fetch:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
